Move log redaction back into base
