+++ base/logging.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/couchbase/clog"
 	"github.com/couchbase/goutils/logging"
-	l "github.com/couchbase/sync_gateway/log"
 	"github.com/natefinch/lumberjack"
 )
 
@@ -216,7 +215,7 @@ type LogAppenderConfig struct {
 	LogKeys        []string           `json:",omitempty"` // Log keywords to enable
 	LogLevel       Level              `json:",omitempty"`
 	Rotation       *LogRotationConfig `json:",omitempty"`
-	RedactionLevel l.RedactionLevel   `json:",omitempty"`
+	RedactionLevel RedactionLevel     `json:",omitempty"`
 }
 
 type LoggingConfigMap map[string]*LogAppenderConfig
@@ -238,16 +237,16 @@ func SetLogLevel(level int) {
 	logLevel = level
 }
 
-func SetRedaction(redactionLevel l.RedactionLevel) {
+func SetRedaction(redactionLevel RedactionLevel) {
 	Logf("Log redaction level: %s", redactionLevel)
 	switch redactionLevel {
-	case l.RedactFull:
-		l.RedactUserData = true
-	case l.RedactPartial:
-		l.RedactUserData = true
-	case l.RedactNone:
+	case RedactFull:
+		RedactUserData = true
+	case RedactPartial:
+		RedactUserData = true
+	case RedactNone:
 	default:
-		l.RedactUserData = false
+		RedactUserData = false
 	}
 }
 
@@ -454,7 +453,7 @@ func LogTo(key string, format string, args ...interface{}) {
 
 // LogToR redacts any arguments implementing the Redactor interface before calling LogTo
 func LogToR(key, format string, args ...interface{}) {
-	LogTo(key, format, l.Redact(args)...)
+	LogTo(key, format, Redact(args)...)
 }
 
 func EnableLogKey(key string) {
@@ -505,7 +504,7 @@ func Logf(format string, args ...interface{}) {
 
 // LogfR redacts any arguments implementing the Redactor interface before calling Logf
 func LogfR(format string, args ...interface{}) {
-	Logf(format, l.Redact(args)...)
+	Logf(format, Redact(args)...)
 }
 
 // If the error is not nil, logs its description and the name of the calling function.
@@ -536,7 +535,7 @@ func Warn(format string, args ...interface{}) {
 
 // WarnR redacts any arguments implementing the Redactor interface before calling Warn
 func WarnR(format string, args ...interface{}) {
-	Warn(format, l.Redact(args)...)
+	Warn(format, Redact(args)...)
 }
 
 // Logs a highlighted message prefixed with "TEMP". This function is intended for
@@ -554,7 +553,7 @@ func LogPanic(format string, args ...interface{}) {
 
 // LogPanicR redacts any arguments implementing the Redactor interface before calling LogPanic
 func LogPanicR(format string, args ...interface{}) {
-	LogPanic(format, l.Redact(args)...)
+	LogPanic(format, Redact(args)...)
 }
 
 // Logs a warning to the console, then exits the process.
@@ -565,7 +564,7 @@ func LogFatal(format string, args ...interface{}) {
 
 // LogFatalR redacts any arguments implementing the Redactor interface before calling LogFatal
 func LogFatalR(format string, args ...interface{}) {
-	LogFatal(format, l.Redact(args)...)
+	LogFatal(format, Redact(args)...)
 }
 
 func logWithCaller(color string, prefix string, format string, args ...interface{}) {
