+++ base/logging_test.go
@@ -15,7 +15,6 @@ import (
 	"log"
 	"testing"
 
-	l "github.com/couchbase/sync_gateway/log"
 	"github.com/couchbaselabs/go.assert"
 )
 
@@ -33,10 +32,10 @@ func assertLogContains(t *testing.T, s string, f func()) {
 }
 
 func TestRedactedLogFuncs(t *testing.T) {
-	username := l.UD("alice")
+	username := UD("alice")
 
-	l.RedactUserData = true
-	defer func() { l.RedactUserData = false }()
+	RedactUserData = true
+	defer func() { RedactUserData = false }()
 	EnableLogKey("TEST")
 
 	assertLogContains(t, "TEST: alice", func() { LogTo("TEST", "%s", username) })
