+++ base/redaction_level.go
@@ -1,4 +1,4 @@
-package log
+package base
 
 import (
 	"errors"
