+++ base/redactor_userdata_test.go
@@ -1,4 +1,4 @@
-package log
+package base
 
 import (
 	"math/big"
